+++ test/db/isoexpress/rt_lmp_hourly_test.dart
@@ -7,16 +7,21 @@ import 'package:timezone/data/latest.dart';
 import 'package:timezone/standalone.dart';
 import 'package:date/date.dart';
 import 'package:elec_server/src/db/isoexpress/rt_lmp_hourly.dart';
+import 'package:dotenv/dotenv.dart' as dotenv;
 
 /// prepare data by downloading a few reports
-prepareData() async {
+Future<void> prepareData() async {
   var archive = RtLmpHourlyArchive();
-  var days = [Date.utc(2017, 3, 12), Date.utc(2017, 11, 5)];
+  var days = [
+    Date.utc(2017, 3, 12),
+    Date.utc(2017, 11, 5),
+    Date.utc(2022, 12, 22),
+  ];
   await archive.downloadDays(days);
 }
 
-RtLmpHourlyTest() async {
-  group('RT hourly lmp report', () {
+Future<void> tests() async {
+  group('ISONE RT hourly lmp report', () {
     late RtLmpHourlyArchive archive;
     setUp(() async {
       archive = RtLmpHourlyArchive();
@@ -25,7 +30,13 @@ RtLmpHourlyTest() async {
     tearDown(() async {
       await archive.dbConfig.db.close();
     });
-
+   test('RT hourly lmp report, DST day spring', () async {
+     var file = archive.getFilename(Date.utc(2022, 12, 22));
+     var res = archive.processFile(file);
+     expect(res.first.keys.toSet(), {'date', 'ptid', 'congestion', 'lmp', 'marginal_loss'});
+     var x321 = res.firstWhere((e) => e['ptid'] == 321);
+     expect(x321['lmp'].first, 136.57);
+   });
 //    test('RT hourly lmp report, DST day spring', () async {
 //      File file = archive.getFilename(new Date.utc(2017, 3, 12));
 //      var res = await archive.processFile(file);
@@ -36,23 +47,23 @@ RtLmpHourlyTest() async {
 //      var res = await archive.processFile(file);
 //      expect(res.first['hourBeginning'].length, 25);
 //    });
-    test('Insert one day', () async {
-      await archive.downloadDay(Date.utc(2017, 1, 1));
-      await archive.insertDay(Date.utc(2017, 1, 1));
-    });
-    test('insert several days', () async {
-      List days =
-          Interval(Date.utc(2017, 1, 1).start, Date.utc(2017, 1, 5).start)
-              .splitLeft((dt) => Date.utc(dt.year, dt.month, dt.day));
-      for (var day in days) {
-        await archive.downloadDay(day);
-        await archive.insertDay(day);
-      }
-    });
+//     test('Insert one day', () async {
+//       await archive.downloadDay(Date.utc(2017, 1, 1));
+//       await archive.insertDay(Date.utc(2017, 1, 1));
+//     });
+//     test('insert several days', () async {
+//       List days =
+//           Interval(Date.utc(2017, 1, 1).start, Date.utc(2017, 1, 5).start)
+//               .splitLeft((dt) => Date.utc(dt.year, dt.month, dt.day));
+//       for (var day in days) {
+//         await archive.downloadDay(day);
+//         await archive.insertDay(day);
+//       }
+//     });
   });
 }
 
-Future fillDb() async {
+Future<void> fillDb() async {
   var archive = RtLmpHourlyArchive();
   await archive.dbConfig.db.open();
   List days = Interval(Date.utc(2017, 12, 31).start, Date.utc(2018, 1, 1).start)
@@ -64,15 +75,12 @@ Future fillDb() async {
   await archive.dbConfig.db.close();
 }
 
-main() async {
+Future<void> main() async {
   initializeTimeZones();
-  // await new RtLmpHourlyArchive().setupDb();
+  dotenv.load('.env/prod.env');
+
+  // await RtLmpHourlyArchive().setupDb();
   // await prepareData();
+  await tests();
 
-  await fillDb();
-
-//  Db db = new Db('mongodb://localhost/isoexpress');
-//  await new DaLmpHourlyArchive().updateDb(new DaLmp(db));
-
-  //await soloTest();
 }